pyproject.toml
[project]
name = "fly-fishing-meniscus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "imageio>=2.37.0",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "opencv-contrib-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "polars>=1.32.0",
    "scipy>=1.16.1",
]
